+++ src/web/Sitemap.php
@@ -6,6 +6,7 @@ use davidhirtz\yii2\skeleton\behaviors\SitemapBehavior;
 use davidhirtz\yii2\skeleton\db\ActiveRecord;
 use davidhirtz\yii2\skeleton\helpers\ArrayHelper;
 use davidhirtz\yii2\skeleton\helpers\FileHelper;
+use davidhirtz\yii2\skeleton\models\interfaces\SitemapInterface;
 use Yii;
 use yii\base\Component;
 use yii\caching\Cache;
@@ -33,8 +34,8 @@ class Sitemap extends Component
     public mixed $variations = null;
 
     /**
-     * @var bool whether sitemaps should be split into separate sitemap files. This is needed for a sitemaps which would
-     * exceed 50 MB or 50.000 URLs.
+     * @var bool whether sitemaps should be split into separate sitemap files. This is necessary for a sitemaps which
+     * would exceed 50 MB or 50.000 URLs.
      */
     public bool $useSitemapIndex = false;
 
@@ -52,7 +53,7 @@ class Sitemap extends Component
     public array $views = [];
 
     /**
-     * @var SitemapBehavior[]|array containing the class definitions of all {@see ActiveRecord} which should be used to
+     * @var SitemapInterface[]|array containing the class definitions of all {@see ActiveRecord} which should be used to
      * create sitemap URLs via the {@see SitemapBehavior} behavior. If `useSitemapIndex` is set to true, the key can
      * optionally be set to a string which is then used in the sitemap index url generation.
      */