+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.3.10 (under development)
 
+- Added `AdminButton::$adminLinkZIndex` (still defaults to '3')
 - Fixed `GridView::$searchUrl` to reset current page on default
 
 ## 2.3.9 (Jul 26, 2024)